+++ docs/CHANGELOG.md
@@ -14,6 +14,12 @@ nav_order: 6
 
     *Adam Hess*, *Joel Hawksley*
 
+## 4.0.0.alpha3
+
+* BREAKING: Remove dependency on `ActionView::Base`, eliminating the need for capture compatibility patch.
+
+    *Cameron Dutro*
+
 ## 4.0.0.alpha2
 
 * Add `#current_template` accessor and `Template#path` for diagnostic usage.